Add options to specify pre-approve, confirm and verify in mass-subscribe.
This adds the ability to control options when mass-subscribing emails to a
MailingList which can be useful.

It also makes it such that the default for pre_verify is now "False" instead of
"True" so that users always verify their email addresses in case it is not
verified in the Mailman Core.