Log all unhandled HTTPError in Postorius.
This basically add an exception handler for all unhandled HTTPError
so that the are logged properly instead of the mysterious `KeyError`
which shows up.

The cause is often an error which wasn't handled in the view where it should
have been handled for the more descriptive error message. This is nonetheless
more useful than the `KeyError` we often see.